Se renombran carpetas se agregan form de buscar y se linkea ruta del producto
