+++ src/app/(paginaPrincipal)/Categorias/cupcakes/page.jsx
@@ -62,11 +62,51 @@ export default function Cupcakes() {
           ></path>
         </svg>
       </div>
+      <div className="grid justify-items-end">
+        <form className="w-[20rem]    md:w-[30rem]  order-2 ">
+          <label className="mb-2 text-sm font-medium text-gray-900 sr-only dark:text-white">
+            Search
+          </label>
+
+          <div className="relative">
+            <div className="absolute inset-y-0 start-0 flex items-center ps-3 pointer-events-none">
+              <svg
+                className="w-4 h-4 text-gray-500 dark:text-gray-400"
+                aria-hidden="true"
+                xmlns="http://www.w3.org/2000/svg"
+                fill="none"
+                viewBox="0 0 20 20"
+              >
+                <path
+                  stroke="currentColor"
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                  strokeWidth="2"
+                  d="m19 19-4-4m0-7A7 7 0 1 1 1 8a7 7 0 0 1 14 0Z"
+                />
+              </svg>
+            </div>
+            <input
+              type="search"
+              id="default-search"
+              className="block w-full p-4 ps-10 text-sm text-gray-900 border border-gray-300 rounded-lg bg-gray-50 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
+              placeholder="Nombre del producto, sabor..."
+              required
+            />
+            <button
+              type="submit"
+              className="text-white absolute end-2.5 bottom-2.5 bg-button_primary hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-4 py-2 dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800"
+            >
+              Buscar
+            </button>
+          </div>
+        </form>
+      </div>
       <div className="pt-50 grid justify-items-center md:grid-cols-3">
         {/* card1 */}
         <div className={opensans.className}>
           <Card className="max-w-xs my-4">
-            <a href="../Categorias/Producto">
+            <a href="../Categories/Product">
               <Image
                 src={cupcake1}
                 alt="cupcake1"
@@ -91,7 +131,7 @@ export default function Cupcakes() {
         {/* card2 */}
         <div>
           <Card className="max-w-xs my-4 ">
-            <a href="../Categorias/Producto">
+            <a href="../Categories/Product">
               <Image
                 src={cupcake2}
                 alt="cupcake2"
@@ -116,7 +156,7 @@ export default function Cupcakes() {
         {/* card3 */}
         <div>
           <Card className="max-w-xs my-4">
-            <a href="../Categorias/Producto">
+            <a href="../Categories/Product">
               <Image
                 src={cupcake3}
                 alt="cupcake3"
@@ -141,7 +181,7 @@ export default function Cupcakes() {
         {/* card4 */}
         <div>
           <Card className="max-w-xs my-4">
-            <a href="../Categorias/Producto">
+            <a href="../Categories/Product">
               <Image
                 src={cupcake4}
                 alt="cupcake4"
@@ -166,7 +206,7 @@ export default function Cupcakes() {
         {/* card5 */}
         <div>
           <Card className="max-w-xs my-4">
-            <a href="../Categorias/Producto">
+            <a href="../Categories/Product">
               <Image
                 src={cupcake5}
                 alt="cupcake5"
@@ -191,7 +231,7 @@ export default function Cupcakes() {
         {/* card6 */}
         <div>
           <Card className="max-w-xs my-4">
-            <a href="../Categorias/Producto">
+            <a href="../Categories/Product">
               <Image
                 src={cupcake6}
                 alt="cupcake6"
