+++ src/app/components/Counter/styles.css
@@ -12,7 +12,7 @@
   height: 50px;
   font-weight: 600;
   font-size: 1.4rem;
-  margin: 40px auto 0;
+  margin: 20px auto 0;
 }
 
 .count-container__button {
